+++ docs/release-notes.md
@@ -1,5 +1,6 @@
 ## 5.2.0
 
+- General: UI can be hidden when launching a custom browser (useful for terminal commands).
 - Toast: notification shows browser icon instead of BT logo.
 - Toast: visibility interval is configurable in settings.
 - Picker: added option to customise window border width (previously there was no border, now will default to 1).